add styles for table
